UIntやIntで値段を扱っていた部分をLongに変更

+++ composeApp/src/commonMain/kotlin/com/punyo/casherapp/ui/product/ProductScreenViewModel.kt
@@ -66,13 +66,13 @@ class ProductScreenViewModel(
     fun addProduct(
         name: String,
         barcode: String?,
-        price: UInt,
+        price: Long,
     ) {
         viewModelScope.launch {
             repository.insertProduct(
                 name = name,
                 barcode = barcode,
-                price = price.toInt().toMoney(),
+                price = price.toMoney(),
             )
         }
     }
@@ -81,14 +81,14 @@ class ProductScreenViewModel(
         id: String,
         name: String,
         barcode: String?,
-        price: UInt,
+        price: Long,
     ) {
         viewModelScope.launch {
             val updatedProduct = ProductDataModel(
                 id = id,
                 name = name,
                 barcode = barcode,
-                price = price.toInt().toMoney(),
+                price = price.toMoney(),
             )
             repository.updateProduct(updatedProduct)
         }
